Улучшена функция отрисовки событий
